// packages/spacelift-io-shared/src/services/Spacelift/types.ts
import { LoggerService } from '@backstage/backend-plugin-api';
import { Stack, TriggerRunResponse } from '../../types';

export type SpaceliftServiceCtx = {
  api: {
    hostUrl: string;
    apiKey: string;
    apiSecret: string;
  };
  logger: LoggerService;
};

export type SpaceliftService = {
  getStacks: () => Promise<Stack[]>;
  triggerRun: (stackId: string) => Promise<TriggerRunResponse>;
  createStack: (stack: Stack) => Promise<Stack>;
};